refactor: organize extraction tests to match source code structure
- Create NativebrikTests/data/extraction/ directory to mirror source structure
- Move all extraction-related tests from experiment-test.swift to extractionTests.swift
- This improves test organization and makes it easier to find corresponding tests
